refactor and split up runner
This allows to only run HTML or CSS stuff only
