+++ lib/headhunter/runners/html_runner.rb
@@ -1,0 +1,17 @@
+module Headhunter
+  class HtmlRunner
+    def initialize
+      @html_validator = HtmlValidator.new
+    end
+
+    def process(url, html)
+      @html_validator.validate(url, html)
+    end
+
+    def results
+      @html_validator.statistics
+    end
+
+    def clean_up ; end
+  end
+end
